# Allow [String: [String]] dictionaries to be resultant of parsing URL-encoded strings