--- a/Tests/KituraTests/TestResponse.swift
+++ b/Tests/KituraTests/TestResponse.swift
@@ -309,6 +309,24 @@ class TestResponse: KituraTest {
                 req.write(from: "swift=rocks")
             }
         }
+        // Now try the multi-value body parser
+        performServerTest(router) { expectation in
+            self.performRequest("post", path: "/bodytestMultiValue", callback: {response in
+                XCTAssertNotNil(response, "ERROR!!! ClientRequest response object was nil")
+                XCTAssertNotNil(response?.headers["Date"], "There was No Date header in the response")
+                do {
+                    let body = try response?.readString()
+                    XCTAssertEqual(body, "<!DOCTYPE html><html><body><b>Received URL encoded body</b><br> [\"swift\": [\"rocks\", \"rules\"]] </body></html>\n\n")
+                } catch {
+                    XCTFail("Error reading body")
+                }
+                expectation.fulfill()
+            }) {req in
+                req.headers["Content-Type"] = "application/x-www-form-urlencoded"
+                req.write(from: "swift=rocks&swift=rules")
+            }
+        }
+
     }
 
     func dataComponentsTest(_ searchString: String, separator: String) {
@@ -1207,6 +1225,13 @@ class TestResponse: KituraTest {
                 } catch {
                     XCTFail("caught error: \(error)")
                 }
+            } else if let urlEncoded = requestBody.asURLEncodedMultiValue {
+                do {
+                    response.headers["Content-Type"] = "text/html; charset=utf-8"
+                    try response.send("<!DOCTYPE html><html><body><b>Received URL encoded body</b><br> \(urlEncoded) </body></html>\n\n").end()
+                } catch {
+                    XCTFail("caught error: \(error)")
+                }
             } else if let text = requestBody.asText {
                 do {
                     response.headers["Content-Type"] = "text/html; charset=utf-8"
@@ -1233,7 +1258,10 @@ class TestResponse: KituraTest {
             next()
         }
 
-        router.all("/bodytest", middleware: BodyParser())
+        router.all("/bodytestMultiValue", allowPartialMatch: false, middleware: BodyParserMultiValue())
+        router.post("/bodytestMultiValue", handler: bodyTestHandler)
+
+        router.all("/bodytest", allowPartialMatch: false, middleware: BodyParser())
         router.post("/bodytest", handler: bodyTestHandler)
 
         router.post("/bodytesthardway") { request, response, next in