…rather than [String: String] dictionaries where multiple values for keys are comma-separated.

<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
A urlDecodedFieldMultiValuePairs() method is added to String alongside the existing urlDecodedFieldValuePairs(). Additionally, a BodyParserMultiValue class is added as a subclass of the existing BodyParser class. (Both of these are kind of ugly since the existing code isn't written to be flexible in this regard…)

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
I had a PR at https://github.com/IBM-Swift/Kitura/pull/1237 that would replace existing code related to decoding URL-encoding GET query strings and POST bodies so that `foo=bar&foo=baz` would result in `["foo": ["bar", "baz"]]` rather than `["foo": "bar,baz"]` as is the current behavior. As this would break pretty much all existing sites that handle such things, @ianpartridge asked me to try again in a way that wouldn't cause breakage. This is the result.

From the other PR:

> The code that decodes the URL encoding to create these dictionaries is in [String+Extensions.swift](https://github.com/IBM-Swift/Kitura/blob/master/Sources/Kitura/String%2BExtensions.swift).
> 
> Consider a form such as: 
> 
> ```html
> <h1>What are your favorite pizza toppings?</h1>
> <form method="get">
>   <label><input type="checkbox" name="toppings" value="extraCheese" />Extra cheese</label>
>   <label><input type="checkbox" name="toppings" value="pepperoni" />Pepperoni</label>
>   <label><input type="checkbox" name="toppings" value="pineapple" />Pineapple</label>
>   <label><input type="checkbox" name="toppings" value="anchovies" />Anchovies</label>
>   <input type="submit" />
> </form>
> ```
> 
> If you were to select both "Extra cheese" and "Pineapple," the resulting URL-encoded string would look something like `toppings=extraCheese&toppings=pineapple`. So there are two values for the same key.
> 
> However, Kitura will only allow a single string value for each key. So Kitura works around this by using a comma to concatenate all of the values. The result is that `request.queryParameters["toppings"]` will be `"extraCheese,pineapple"`.
> 
> I don't like this. For one, it makes it impossible to simply count how many values you have for a key without doing something like `.split(separator: ",").count`. For another, what if a value itself has a comma? Now we're in painful territory.

With this PR, `request.queryParametersMultiValues` can be used in place of `request.queryParameters` to get values for a key as a `[String]`. Similarly, for POST forms, the following can be done:

```swift
router.post("/submit", middleware: BodyParserMultiValue()) // instead of BodyParser()
router.post("/submit") { request, response, next in
    if let postValues = request.body?.asURLEncodedMultiValue { // instead of .asURLEncoded
        // postValues is a [String: [String]] now
```


## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Test cases have been added for both GET and POST requests.

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] If applicable, I have updated the documentation accordingly.
- [x] If applicable, I have added tests to cover my changes.