@NocturnalSolutions 2018-04-13
Pushed tweaks to avoid force unwrapping as per @djones6 's concerns. (Thanks to ch1ck3n in #swift-lang (Freenode) for reminding me of the `default` trick.) 

Perf tested again under similar circumstances:

```
> wrk -t8 -c100 -d1m 'http://localhost:8080/?foo=bar&baz=qux&multival=123&multival=234'
Running 1m test @ http://localhost:8080/?foo=bar&baz=qux&multival=123&multival=234
  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     8.57ms    6.65ms 191.45ms   97.70%
    Req/Sec     1.50k   107.23     2.29k    74.62%
  715253 requests in 1.00m, 137.79MB read
Requests/sec:  11907.66
Transfer/sec:      2.29MB
```

Huh, so that's actually faster than my test with the unpatched code base, which doesn't make sense unless the speed change is actually negligible and all these results are just falling within a margin of error.