@djones6 2018-04-16
@NocturnalSolutions There is a fair margin of error with this sort of testing - your best and worst results are 2% apart, which I'd say is inconclusive based on a single execution of each test.

A way to get around that is to execute each test multiple times (before, after, before, after...) and compare the average of the befores with the average of the afters.

The test itself that you are performing looks fine, but I'd say an equally interesting case would be where you have no multi values at all:  the new code creates an empty array, appends a single element, then invokes `.joined` on the single-element array.  (I don't know what degree of fast paths the stdlib / compiler apply in this case).

I'd also expect that the level of regression a change like this could cause to a Kitura workload is very small - this is small fry considering the total cost of reading/writing to the socket, constructing the Request, Response objects, HTTP parsing, headers container and all the other bits and bobs involved in serving a request.

If we want to measure which approach is faster, for the sake of interest, then you could extract the before/after code into a standalone testcase and time it running in a loop.

Oh, one other (crucial) detail: make sure you compile with `swift build -c release` (or `swiftc -O` for simple scripts) when comparing performance: release mode enables compiler optimizations that are not applied to the default (debug) build, which can dramatically affect a comparison.