@NocturnalSolutions 2018-04-16
Okay, here's some results from going back-and-forth between patched and unpatched packages and testing in between, and using that `-c release` bit (which made a rather significant difference). The command run this time was `http://localhost:8080/?foo=bar&baz=qux&multival=123` - so no actual multiple value GET parameter this time. Also, unlike my previous testing, I didn't quit all my other Mac apps besides Terminal this time.

Before results:

```
Running 1m test @ http://localhost:8080/?foo=bar&baz=qux&multival=123
  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.15ms   12.03ms 230.47ms   94.67%
    Req/Sec     2.44k   319.24     4.05k    78.81%
  1164897 requests in 1.00m, 219.96MB read
Requests/sec:  19390.12
Transfer/sec:      3.66MB

Running 1m test @ http://localhost:8080/?foo=bar&baz=qux&multival=123
  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.69ms   13.80ms 255.25ms   94.61%
    Req/Sec     2.37k   371.07     4.00k    84.33%
  1134102 requests in 1.00m, 214.15MB read
Requests/sec:  18876.50
Transfer/sec:      3.56MB

  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.81ms   14.81ms 234.57ms   94.50%
    Req/Sec     2.45k   298.84     4.76k    76.06%
  1169577 requests in 1.00m, 220.85MB read
Requests/sec:  19468.90
Transfer/sec:      3.68MB
```

Average req/sec: 19,245.17

After patching:

```
Running 1m test @ http://localhost:8080/?foo=bar&baz=qux&multival=123
  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.38ms   12.23ms 235.99ms   95.33%
    Req/Sec     2.25k   316.03     6.98k    82.35%
  1074179 requests in 1.00m, 202.83MB read
Requests/sec:  17883.98
Transfer/sec:      3.38MB

Running 1m test @ http://localhost:8080/?foo=bar&baz=qux&multival=123
  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.86ms   10.57ms 224.09ms   95.64%
    Req/Sec     2.31k   310.92     3.68k    81.40%
  1102305 requests in 1.00m, 208.15MB read
Requests/sec:  18356.73
Transfer/sec:      3.47MB

Running 1m test @ http://localhost:8080/?foo=bar&baz=qux&multival=123
  8 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.89ms   13.85ms 225.28ms   94.38%
    Req/Sec     2.33k   304.64     4.25k    77.88%
  1113263 requests in 1.00m, 210.21MB read
Requests/sec:  18535.43
Transfer/sec:      3.50MB
```

Average req/sec: 18,258.71

Just as an aside, I know this is an artificial test with no real network use, but still, over 18 thousand requests per second on my oldish MBP? That's an unimaginably big number to me, and I don't think I could get similar numbers with a PHP-based stack on the same hardware without a lot more work. Great job, Swift and Kitura teams!