@djones6 2018-04-24
I ran some measurements on a Linux machine locally with the same route as @NocturnalSolutions - there is regression, the amount of regression depends on the query params specified but is surprising:

- No query params:  no regression (70.8k vs. 70.9k)
- `?a=b`:  2% regression (43.9k vs. 44.8k)
- `?a=b&c=d&a=c&a=d`:  3% regression (30.3k vs. 31.3k)
- `?a=b&a=c&a=d&a=e&a=f&a=g&a=h&a=i&a=j&a=k`:  1.5% regression (21k vs. 21.3k)
- `?a=b&c=d&e=f&g=h&i=j&k=l&m=n&o=p&q=r&s=t`: 6% regression (17.6k vs. 18.7k)

The regression increases with the number of distinct query params:  a single query param with 10 values shows only 1.5% regression, whereas 10 query params each with a single value regresses by 6%.

I'm particularly surprised that this code could cause such a significant regression, but it does indicate that going via `urlDecodedFieldMultiValuePairs` is substantially more expensive for single value params than the existing code.  As a quick experiment, I tried replacing the `values.joined(separator:)` with `values.first` but this made no measurable difference to the 6% case above,  so it seems the whole cost is in building the intermediate array representation while computing `urlDecodedFieldMultiValuePairs`.

For this reason I think it would be worth the extra bit of code duplication to maintain the two as independent functions to avoid the regression for the common case (single value params).