@NocturnalSolutions 2018-04-24
Okay, pushed a change to duplicate the code in the relevant part. (I also gave up on trying to get the encoding/decoding bit in KituraContracts to play along. I don't think I had that working 100% in the first place. At any rate, I guess it can be saved for another day. So now you don't have to also patch KituraContracts to test this stuff.) I think there's deduplication that can be done in other ways, but for now, at commit 2e119e3, with `wrk -t8 -c100 -d1m 'http://localhost:8080/?foo=bar&baz=qux&multival=123&a=b&c=d&e=f&g=h&i=j&k=l&mno=pqrst' `, I get 10,841.50 reqs/sec before the patch and 11,060.16 reqs/sec after - probably due to statistical noise.