@djones6 2018-05-01
I confirmed on Linux that this is now performance-neutral, so from that perspective I think it is fine to go ahead.